backend/requirements-ci.txt
fastapi==0.133.1
uvicorn[standard]==0.41.0
sqlalchemy==2.0.47
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
email-validator==2.3.0
PyJWT==2.12.1
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.18.4
supabase>=2.0.0
pytest==9.0.2
pytest-cov==6.1.1
httpx==0.28.1
anyio==4.12.1
ruff==0.11.4
